+++ backend/server-core/server_core/views/js_client_views/project_view.py
@@ -4,7 +4,7 @@ from flask import Blueprint, Response, request
 from peewee import DoesNotExist
 
 from amapy_pluggy.storage import BlobStoreURL
-from asset_plugin_gcs.bucket_cors import update_cors_configuration as gcs_update_cors
+from amapy_plugin_gcs.bucket_cors import update_cors_configuration as gcs_update_cors
 from asset_plugin_s3.bucket_cors import set_bucket_cors as s3_update_cors
 from server_core.configs import Configs
 from server_core.models.project import Project
