+++ client/asset-plugin-gcs/amapy_plugin_gcs/bucket_cors.py
