+++ client/asset-plugin-gcs/amapy_plugin_gcs/gcs_blob.py
