+++ client/asset-plugin-gcs/amapy_plugin_gcs/gcs_storage_mixin.py
@@ -2,7 +2,7 @@ from google.cloud import storage as gcs
 
 from amapy_pluggy.storage import BlobStoreURL
 from amapy_pluggy.storage import storage_utils
-from asset_plugin_gcs.transporter.async_gcs import async_delete
+from amapy_plugin_gcs.transporter.async_gcs import async_delete
 
 
 class GcsStorageMixin:
