+++ client/asset-plugin-gcs/amapy_plugin_gcs/signed_url.py
