+++ client/asset-plugin-gcs/amapy_plugin_gcs/tests/test_gcs_storage.py
@@ -2,8 +2,8 @@ import json
 import os
 from unittest.mock import patch
 
-from asset_plugin_gcs.gcs_blob import GcsBlob
-from asset_plugin_gcs.gcs_storage import GcsStorage
+from amapy_plugin_gcs.gcs_blob import GcsBlob
+from amapy_plugin_gcs.gcs_storage import GcsStorage
 from amapy_utils.utils import list_files
 from amapy_utils.utils.file_utils import FileUtils
