+++ client/asset-plugin-gcs/amapy_plugin_gcs/transporter/async_gcs/async_copy.py
@@ -5,8 +5,8 @@ import json
 import aiohttp
 import backoff
 
-from asset_plugin_gcs.transporter.async_gcs.async_storage import AsyncStorage
-from asset_plugin_gcs.transporter.gcs_transport_resource import GcsCopyResource
+from amapy_plugin_gcs.transporter.async_gcs.async_storage import AsyncStorage
+from amapy_plugin_gcs.transporter.gcs_transport_resource import GcsCopyResource
 from amapy_utils.utils.log_utils import get_logger
 
 logger = get_logger(__name__)
