+++ client/asset-plugin-gcs/amapy_plugin_gcs/transporter/async_gcs/async_gcs_transporter.py
@@ -1,8 +1,8 @@
 import aiohttp
 
 from amapy_pluggy.storage.transporter import Transporter, TransportResource
-from asset_plugin_gcs.transporter.async_gcs import async_upload, async_download, async_copy
-from asset_plugin_gcs.transporter.gcs_transport_resource import GcsUploadResource, GcsDownloadResource, GcsCopyResource
+from amapy_plugin_gcs.transporter.async_gcs import async_upload, async_download, async_copy
+from amapy_plugin_gcs.transporter.gcs_transport_resource import GcsUploadResource, GcsDownloadResource, GcsCopyResource
 from amapy_utils.common import exceptions
 from amapy_utils.utils import utils
