+++ client/asset-plugin-gcs/amapy_plugin_gcs/transporter/async_gcs/async_upload.py
@@ -6,8 +6,8 @@ import os
 import aiohttp
 import backoff
 
-from asset_plugin_gcs.transporter.async_gcs.async_storage import AsyncStorage
-from asset_plugin_gcs.transporter.gcs_transport_resource import GcsUploadResource
+from amapy_plugin_gcs.transporter.async_gcs.async_storage import AsyncStorage
+from amapy_plugin_gcs.transporter.gcs_transport_resource import GcsUploadResource
 from amapy_utils.utils.log_utils import get_logger
 
 logger = get_logger(__name__)
