+++ client/asset-plugin-gcs/amapy_plugin_gcs/transporter/gcs_transport_resource.py
